UPSTREAM: crypto: arm64/sha2-ce - fix for big endian
The SHA256 digest is an array of 8 32-bit quantities, so we should refer
to them as such in order for this code to work correctly when built for
big endian. So replace 16 byte scalar loads and stores with 4x32 vector
ones where appropriate.

Fixes: 6ba6c74dfc6b ("arm64/crypto: SHA-224/SHA-256 using ARMv8 Crypto Extensions")
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
(cherry picked from commit 174122c39c369ed924d2608fc0be0171997ce800)
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>
Signed-off-by: Ícaro Hoff <icarohoff@gmail.com>
Signed-off-by: Lucchetto <lucchetto.tie@live.com>