Kconfig, power, usb: adjust LE_AB framework and fix HVDCP3
After a bit of debugging and comparison with the latest stock Kernel
available from OEM, I've finally managed to get the charger code working
almost perfectly.

The main reason it didn't work was due to the OEM Open Kernel having the
order for LE_AB wrong.

In this change we'll beat those two issues by:

1. Making LE_AB framework optional through PD configuration.
2. Adjusting voltage measurements to be independent from LE_AB.
3. Adjusting 'Analogix Ohio' USB configuration dependency on LE_PD.

Change-Id: Idb7a40ad3c57868329be105889454f2b9ac6f5a4
Signed-off-by: Ícaro Hoff <icarohoff@gmail.com>